Revert "Revert "Refactor `shopify hydrogen deploy` into separate folders""
