+++ packages/cli-hydrogen/src/cli/services/build.ts
@@ -7,9 +7,10 @@ interface DevOptions {
   directory: string
   targets: {[key in Target]: boolean | string}
   base?: string
+  assetBaseURL?: string
 }
 
-async function build({directory, targets, base}: DevOptions) {
+async function build({directory, targets, base, assetBaseURL}: DevOptions) {
   const commonConfig = {base, root: directory}
 
   const tasks: ui.ListrTask[] = Object.entries(targets)
@@ -21,6 +22,10 @@ async function build({directory, targets, base}: DevOptions) {
           if (key === 'worker') {
             process.env.WORKER = 'true'
           }
+          if (assetBaseURL) {
+            process.env.HYDROGEN_ASSET_BASE_URL = assetBaseURL
+          }
+
           try {
             await viteBuild({
               ...commonConfig,
