+++ packages/cli-hydrogen/src/cli/services/deploy/graphql/create_deployment.ts
@@ -1,0 +1,26 @@
+import {DMSError} from '../types.js'
+import {gql} from 'graphql-request'
+
+export const CreateDeploymentQuery = gql`
+  mutation createDeployment($input: CreateDeploymentInput!) {
+    createDeployment(input: $input) {
+      deploymentID
+      assetBaseURL
+      error {
+        code
+        unrecoverable
+        debugInfo
+      }
+    }
+  }
+`
+
+export interface CreateDeploymentQuerySchema {
+  createDeployment: CreateDeploymentResponse
+}
+
+export interface CreateDeploymentResponse {
+  deploymentID: string
+  assetBaseURL: string
+  error: DMSError
+}
