+++ packages/cli-hydrogen/src/cli/services/deploy/types.ts
@@ -1,0 +1,18 @@
+export interface DeployConfig {
+  deploymentToken: string
+  dmsAddress: string
+  commitMessage?: string
+  commitAuthor?: string
+  commitSha?: string
+  commitRef?: string
+  timestamp?: string
+  repository?: string
+  path?: string
+}
+export type ReqDeployConfig = Required<DeployConfig>
+
+export interface DMSError {
+  code: string
+  unrecoverable: boolean
+  debugInfo: string
+}
